+++ src/dk/aau/cs/ds306e18/tournament/ui/bracketObjects/SingleEliminationNode.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 /** Used to display the a single elimination stage. */
 public class SingleEliminationNode extends GridPane implements ModelCoupledUI {
 
-    private final Insets MARGINS = new Insets(0, 0, 8, 0);
+    private final Insets MARGINS = new Insets(0, 0, 16, 0);
     private final int CELL_SIZE = 50;
 
     private final SingleEliminationFormat singleElimination;
@@ -54,6 +54,7 @@ public class SingleEliminationNode extends GridPane implements ModelCoupledUI {
                     MatchVisualController mvc = boc.loadVisualMatch(match);
                     mvcs.add(mvc);
                     box.getChildren().add(mvc.getRoot());
+                    mvc.setShowIdentifier(true);
                 }
 
                 box.setAlignment(Pos.CENTER);